LAb12-1 complete yessir

+++ lab12_1.cpp
@@ -2,7 +2,7 @@
 #include <ctime>
 using namespace std;
 
-int fibonacci(int x);
+long long int fibonacci(long long int x);
 
 int main(){
     clock_t start = clock();
@@ -13,7 +13,7 @@ int main(){
     return 0;
 }
 
-int fibonacci(int x){
+long long int fibonacci(long long int x){
     if (x==0){
         return 0;
     }else if (x==1){
